ci/publish-go-docs-as-module-in-registry (#12)
* build: Update GitHub Actions workflows to fetch full history

The release.yaml files in the workflows directory were updated to include
the "fetch-depth: 0" option in the "actions/checkout" step. This change allows
the workflows to fetch the full commit history when checking out the repository
before running the actions. Additionally, the publish-go-module.sh script was
modified to check for the latest tag when extracting the module version, 
providing more accurate version information.

* feat: Add go-publish task to publish Go module

Added a new task in TaskFile.yml to publish the go module to the registry using a script.
Also, added commands in the publish-go-module.sh script to trigger indexing on pkg.go.dev after publishing the module.