feat: Add support for setting tags in APKO builds
This commit adds a new method `WithTag` to the `ApkoBuilder` struct, which allows users to set a tag for the APKO build. If no tag is provided, it defaults to "latest".

The changes include:

- Adding the `tags` field to the `ApkoBuilder` struct to store the tags.
- Implementing the `WithTag` method to set the tags.
- Updating the `BuildCommand` method to include the `--tag` flag with the specified tag(s).
- Adding unit tests to verify the functionality of the `WithTag` method.

These changes provide more flexibility for users to customize the APKO build process and specify the desired tag for the output image.