feat: Add support for setting tags in APKO builds (#55)
* feat: Add support for setting tags in APKO builds

This commit adds a new method `WithTag` to the `ApkoBuilder` struct, which allows users to set a tag for the APKO build. If no tag is provided, it defaults to "latest".

The changes include:

- Adding the `tags` field to the `ApkoBuilder` struct to store the tags.
- Implementing the `WithTag` method to set the tags.
- Updating the `BuildCommand` method to include the `--tag` flag with the specified tag(s).
- Adding unit tests to verify the functionality of the `WithTag` method.

These changes provide more flexibility for users to customize the APKO build process and specify the desired tag for the output image.

* feat: Add handling for empty mount prefix in GetCacheDir

This commit adds support for handling an empty mount prefix in the GetCacheDir function. If the mount prefix is empty, the function will use the default fixtures.MntPrefix value to construct the cache directory path.

The changes include:

- Adding a check for an empty mount prefix in the GetCacheDir function and setting the prefix to the default value if it's empty.
- Adding a new test case to cover the scenario where the mount prefix is empty.

* feat: Implement GetApkoConfigOrPreset function to validate config file paths

Adds a new function GetApkoConfigOrPreset that validates the provided config file path. It checks if the config file is required, has a valid extension, and if the extension is either .yaml or .yml. This function helps ensure that the config file is properly configured before processing.

* feat: Remove unused imports and update tests for GetCacheDir function

The changes in this commit are focused on removing unused imports and updating the tests for the `GetCacheDir` function in the `pkg/builderx/apko_test.go` file.

The key changes are:

1. Removed the unused `github.com/Excoriate/daggerx/pkg/fixtures` import.
2. Simplified the tests for the `GetCacheDir` function by removing the `WithEmptyMntPrefix` test case, as the function now handles the empty `mntPrefix` case correctly.
3. Updated the error message in the `TestApkoBuilder` test to provide more detailed information about the expected and actual command outputs.
4. Removed the `WithTag` tests, as the `WithTag` functionality has been removed from the `ApkoBuilder`.

* feat: Add tests for GetApkoConfigOrPreset function

This commit adds new tests to cover the functionality of the GetApkoConfigOrPreset function. The tests cover various scenarios, including valid config file names, empty mntPrefix, empty config file, config file without extension, and config file with invalid extension. The tests ensure that the function behaves as expected in these cases.

* feat: Add support for multiple CPU architectures in APKO builds

  
This commit introduces support for multiple CPU architectures in APKO builds. It defines
the supported architectures as constants and adds a new `WithBuildArch` method to the
`ApkoBuilder` struct to set the desired build architecture. This will allow users to
build APKO images for different CPU architectures, such as x86_64, aarch64, armv7,
ppc64le, and s390x.

* feat: Add default APKO repository URL constant

* feat: Improve ApkoBuilder configuration options

- Add new struct fields to ApkoBuilder to support additional configuration options
- Add methods to set the build architecture, build context, and extra arguments
- Refactor the existing ApkoBuilder struct definition to be more concise and organized